Minor: just added id validation to the delete route

+++ Simple-Store/backend/controllers/product.controller.js
@@ -50,11 +50,16 @@ const updateProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
   const id = req.params.id;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted successfully" });
   } catch (err) {
-    res.status(404).json({ success: false, message: "Product Not found" });
+    // res.status(404).json({ success: false, message: "Product Not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
